push commit

+++ fundamentals/22_for_loop_variations/ForTest1.java
@@ -1,6 +1,6 @@
 public class ForTest1 {
     public static void main(String[] args) {
-        for(int i = 0;i < 10;i++){
+        for(double i = 0;i < 10;i++){
             System.out.println("Counter = "+i);
         }
     }
